# lugia-backend/queries/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.30.0

package queries

import (
	"net/netip"

	"github.com/jackc/pgx/v5/pgtype"
)

type EmailChangeToken struct {
	ID        pgtype.UUID        `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	NewEmail  string             `json:"new_email"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
}

type InvitationToken struct {
	ID        pgtype.UUID        `json:"id"`
	TenantID  pgtype.UUID        `json:"tenant_id"`
	UserID    pgtype.UUID        `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
}

type IpWhitelistEmergencyToken struct {
	ID        pgtype.UUID        `json:"id"`
	Jti       pgtype.UUID        `json:"jti"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type PasswordResetToken struct {
	ID        pgtype.UUID        `json:"id"`
	UserID    pgtype.UUID        `json:"user_id"`
	TokenHash string             `json:"token_hash"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
	UsedAt    pgtype.Timestamptz `json:"used_at"`
}

type Permission struct {
	ID          pgtype.UUID        `json:"id"`
	Resource    string             `json:"resource"`
	Action      string             `json:"action"`
	Description string             `json:"description"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type RefreshToken struct {
	ID         pgtype.UUID        `json:"id"`
	UserID     pgtype.UUID        `json:"user_id"`
	Jti        pgtype.UUID        `json:"jti"`
	DeviceInfo pgtype.Text        `json:"device_info"`
	IpAddress  pgtype.Text        `json:"ip_address"`
	ExpiresAt  pgtype.Timestamptz `json:"expires_at"`
	CreatedAt  pgtype.Timestamptz `json:"created_at"`
	UsedAt     pgtype.Timestamptz `json:"used_at"`
	RevokedAt  pgtype.Timestamptz `json:"revoked_at"`
}

type Role struct {
	ID          pgtype.UUID        `json:"id"`
	TenantID    pgtype.UUID        `json:"tenant_id"`
	Name        string             `json:"name"`
	Description pgtype.Text        `json:"description"`
	IsDefault   bool               `json:"is_default"`
	CreatedAt   pgtype.Timestamptz `json:"created_at"`
	UpdatedAt   pgtype.Timestamptz `json:"updated_at"`
}

type RolePermission struct {
	RoleID       pgtype.UUID        `json:"role_id"`
	PermissionID pgtype.UUID        `json:"permission_id"`
	TenantID     pgtype.UUID        `json:"tenant_id"`
	CreatedAt    pgtype.Timestamptz `json:"created_at"`
}

type SsoAuthRequest struct {
	RequestID string             `json:"request_id"`
	TenantID  pgtype.UUID        `json:"tenant_id"`
	Email     string             `json:"email"`
	ExpiresAt pgtype.Timestamptz `json:"expires_at"`
}

type Tenant struct {
	ID                 pgtype.UUID        `json:"id"`
	Name               string             `json:"name"`
	EnterpriseFeatures []byte             `json:"enterprise_features"`
	StripeCustomerID   pgtype.Text        `json:"stripe_customer_id"`
	AuthMethod         string             `json:"auth_method"`
	CreatedAt          pgtype.Timestamptz `json:"created_at"`
	UpdatedAt          pgtype.Timestamptz `json:"updated_at"`
}

type TenantIpWhitelist struct {
	ID        pgtype.UUID        `json:"id"`
	TenantID  pgtype.UUID        `json:"tenant_id"`
	IpAddress netip.Prefix       `json:"ip_address"`
	Label     pgtype.Text        `json:"label"`
	CreatedBy pgtype.UUID        `json:"created_by"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}

type User struct {
	ID              pgtype.UUID        `json:"id"`
	TenantID        pgtype.UUID        `json:"tenant_id"`
	Email           string             `json:"email"`
	PasswordHash    string             `json:"password_hash"`
	Name            string             `json:"name"`
	IsInternalAdmin bool               `json:"is_internal_admin"`
	IsInternalUser  bool               `json:"is_internal_user"`
	CreatedAt       pgtype.Timestamptz `json:"created_at"`
	UpdatedAt       pgtype.Timestamptz `json:"updated_at"`
	DeletedAt       pgtype.Timestamptz `json:"deleted_at"`
	Status          string             `json:"status"`
	ExternalSsoID   pgtype.Text        `json:"external_sso_id"`
}

type UserRole struct {
	UserID    pgtype.UUID        `json:"user_id"`
	RoleID    pgtype.UUID        `json:"role_id"`
	TenantID  pgtype.UUID        `json:"tenant_id"`
	CreatedAt pgtype.Timestamptz `json:"created_at"`
}